+++ lib/io/stream/generic.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # Released under the MIT License.
-# Copyright, 2023-2025, by Samuel Williams.
+# Copyright, 2023-2026, by Samuel Williams.
 
 require_relative "string_buffer"
 require_relative "readable"
@@ -9,6 +9,7 @@ require_relative "writable"
 
 require_relative "shim/buffered"
 require_relative "shim/readable"
+require_relative "shim/timeout"
 
 require_relative "openssl"